Fix duplicated entries in Hall of Fame created
Let's take the following scenario: you start a game from a fresh state (Hall of
Fame is clear). You complete the game. Your name appears as the first entry in
the Hall of Fame. Then you
want to go back to some of the already completed levels. After completing a
level, another entry with your name, with the same time is created in the
Hall of Fame (screenshot). This is bec
ause the completedAllLevels field of the PlayerEntry structure was updated
incorrectly. The code in updateHallOfFameEntries function took the PlayerEntry
and made a local copy of it, so the
original PlayerEntry structure was never updated. I changed this, so now a
pointer to a PlayerEntry is taken.

Fixed by Krzysztof Nowaczyk, thank you! :heart:

Original commit: a3c6818f108dc3827b077442ee496695ae099550