Fix demo signatures never being appended
When recording a demo, I found that when the signature is about to be written to
the output file, the code tries to write to a wrong file handle. It tries to
write to the signature file
(which is open in read-only mode) instead of the demo file. One last bug: there
was an "off by one" error when writing the signature to the output file and the
last terminating 0xFF byte was not written to the file.

Fixed by Krzysztof Nowaczyk, thank you! :heart:

Original commit: 5cc59690a0f274ee14054f6893eff593887ad982
I ommited the filename case change, as it behaves as all other files
do; they must be all in uppercase.